Refactor : `AnswerController` -> `ArticleController`로 병합
- DDD를 공부하며 설계에 변화를 주었다. 메인 도메인을 통해 서브 도메인을 제어할 수 있도록 하기 위해 controller를 병합하였다. repository를 기준으로 해야하나 하는 생각이 들었지만, 더 공부 후 고려해볼 예정.
