+++ src/main/java/kr/codesqaud/cafe/controller/ArticleController.java
@@ -1,5 +1,6 @@
 package kr.codesqaud.cafe.controller;
 
+import kr.codesqaud.cafe.domain.Answer;
 import kr.codesqaud.cafe.domain.Article;
 import kr.codesqaud.cafe.domain.Member;
 import kr.codesqaud.cafe.dto.article.ArticleResponse;
@@ -11,12 +12,15 @@ import kr.codesqaud.cafe.repository.AnswerRepository;
 
 import kr.codesqaud.cafe.repository.ArticleRepository;
 import kr.codesqaud.cafe.repository.MemberRepository;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
+import java.sql.SQLException;
 import java.util.List;
 
 
@@ -28,6 +32,8 @@ import static kr.codesqaud.cafe.exception.ManageArticleException.NOT_POSSIBLE_DE
 @Controller
 public class ArticleController {
 
+    private static final String REDIRECT_ARTICLE = "redirect:/articles/{articleId}";
+
     private final ArticleRepository articleRepository;
     private final MemberRepository memberRepository;
     private final AnswerRepository answerRepository;
@@ -36,6 +42,47 @@ public class ArticleController {
         this.articleRepository = articleRepository;
         this.memberRepository = memberRepository;
         this.answerRepository = answerRepository;
+    }
+
+    @PostMapping("/articles/{articleId}/answers")
+    public String saveAnswer(@PathVariable long articleId, Answer answer, HttpSession httpSession, RedirectAttributes redirectAttributes) {
+        SessionUser sessionUser = SessionUser.getSessionUser(httpSession);
+
+        answer.setWriter(new Member(sessionUser.getId(), sessionUser.getNickName()));
+        answer.setArticleId(articleId);
+        answerRepository.save(answer);
+
+        redirectAttributes.addFlashAttribute("articleId", articleId);
+        return REDIRECT_ARTICLE;
+    }
+
+    // TODO 모르고 댓글 수정을 만들었다.... 하지만 뷰가 없다....
+    @PutMapping("/articles/{articleId}/answers/{answerId}")
+    public String updateAnswer(@PathVariable long articleId, @PathVariable long answerId, Answer answer, HttpSession httpSession, RedirectAttributes redirectAttributes) throws ManageArticleException {
+        SessionUser sessionUser = SessionUser.getSessionUser(httpSession);
+        AnswerResponseDto exAnswer = answerRepository.findById(answerId);
+
+        if (!sessionUser.equals(exAnswer.getWriterId())) {
+            throw new ManageArticleException(INVALID_WRITER);
+        }
+
+        answerRepository.update(answerId, answer.getContents());
+        redirectAttributes.addFlashAttribute("articleId", articleId);
+        return REDIRECT_ARTICLE;
+    }
+
+    @DeleteMapping("/articles/{articleId}/answers/{answerId}")
+    public String deleteAnswer(@PathVariable long articleId, @PathVariable long answerId, HttpSession httpSession, RedirectAttributes redirectAttributes) throws ManageArticleException {
+        SessionUser sessionUser = SessionUser.getSessionUser(httpSession);
+        AnswerResponseDto exAnswer = answerRepository.findById(answerId);
+
+        if (!sessionUser.equals(exAnswer.getWriterId())) {
+            throw new ManageArticleException(INVALID_WRITER);
+        }
+
+        answerRepository.delete(answerId);
+        redirectAttributes.addFlashAttribute("articleId", articleId);
+        return REDIRECT_ARTICLE;
     }
 
     @PostMapping("/questions")
@@ -47,8 +94,9 @@ public class ArticleController {
         return REDIRECT_INDEX;
     }
 
+    @Transactional(rollbackFor = Exception.class)
     @DeleteMapping("/articles/{id}")
-    public String deleteArticle(@PathVariable Long id, HttpSession httpSession) throws InvalidAuthorityException, ManageArticleException {
+    public String deleteArticle(@PathVariable Long id, HttpSession httpSession) throws InvalidAuthorityException, ManageArticleException, SQLException {
         ArticleResponse exArticle = articleRepository.findById(id);
         List<AnswerResponseDto> answerList = answerRepository.findAllByArticleId(id);
         SessionUser sessionUser = SessionUser.getSessionUser(httpSession);
@@ -62,7 +110,10 @@ public class ArticleController {
         }
 
         answerRepository.deleteAll(id);
-        articleRepository.delete(id);
+        int deleteCount = articleRepository.delete(id);
+        if (deleteCount!=1) {
+            throw new SQLException("잘못된 삭제 요청입니다.");
+        }
         return REDIRECT_INDEX;
     }
